test(launcher): move the supervisor Solo-gate tests out of the mocked file
`event-queue.test.ts` replaces `yjs/provider.js` wholesale with a two-export
`vi.mock`. On this branch the supervisor additionally imports
`codex/approval-broker.js`, which reaches `mcp/annotations.js` →
`mcp/document.js` → `documents/registry.js`, and that calls
`setShouldKeepDocument` on the mock at module scope. Importing the supervisor
into that file therefore fails the whole FILE at load — surfaced by vitest as
`1 failed | 403 passed` with zero failing tests, which points nowhere near the
cause. It only appeared once #1265 and #1267 were reunified: neither branch's
import graph reaches the mock on its own.

Moved rather than papered over by widening the mock, which is what the file's
own note asks for ("If that changes, move CTRL_ROOM tests to a separate file").
The new file uses the real provider and real queue.

Adds the indeterminate-mode case while relocating: the gate tests `=== "tandem"`
rather than `!== "solo"`, so an absent mode key must fail closed. Negative
control: flipping `"external"` to `"internal"` reds both withholding tests and
leaves the Tandem positive control green.
